Added function to create moving zone around an entity
CreateAroundEntity uses the model dimensions of an entity to create a zone around it that follows its rotation and position. This zone can be scaled and offset as well.

Entity zones don't use the grid optimization because they are always rectangles, and therefore the winding number algorithm is fast enough.

This also allows us to completely avoid the startup cost of the grid creation.