Added custom pointInPoly check for entity zones
I realized that because all entity zones are rectangular, a bounding box check is all you need to test point inclusion, as long as you properly rotate and offset the point. So entity zones no longer use _pointInPoly, but instead just rotate and offset the point and then do a bounding box check.

This is more efficient and allows us to not make the _pointInPoly code messier than it needs to be.