Implemented realtime Z calculation with entity zones
The minZ/maxZ were calculated as the longest dimension of the entity. Now they are update realtime and exactly match the min and max Z values of the entity dimensions. This is slightly more expensive, but significantly more accurate.

This also allowed changing startPos and offsetPos to vector2s and avoided some cloning of vectors that offset some of the performance regression.