Added new zone type: ComboZone
A ComboZone is a zone of many zones. To create one, you pass in a table of zones and all the functions of zones like isPointInside() and onPointInOut operate on all the zones. This allows sharing duplicate work, like onPointInOut getting the point to test only once instead of many times.