Redo creation to work with poly, box, and circle zones
This is a complete redo of creation. It adds some useful things like text boxes instead of needing to remember command ordering, the ability to start creation of the last circle or box zone you finished creating, and some interactive controls for box and circle zones.

When creating a circle zone, you can use the scroll wheel to change the radius, and ctrl+scroll wheel to change it by a smaller amount

When creating a box zone, you can use the scroll wheel to change the heading/rotation of the zone, ctrl+scroll wheel to change it by a smaller amount, shift+scroll wheel to change the length, alt+scroll wheel to change the length, and the arrow keys to change the center

The command /pzlast will start creation with the same circle/box zone you last finished creating, except it will start in your current position