Add arrow input to circle and poly zone creation
Add ability to use arrow keys to move circle and poly zones during creation. The arrow keys will move the center, in the case of a circle zone, and the last placed point, in the case of a poly zone. You can hold control down to move slower.

Also, added useZ toggle to circle zone creation.