ComboZone: Added enteredZones and leftZones parameters to onPointInOutCb for exhaustive zones
Regular onPointInOut is nice on ComboZones because you can determine whether you entered or left the zone that triggered the callback using the isInside parameter. onPointInOutExhaustive lost this functionality because all it passed back was a table of the zones you were currently inside. To add back this useful functionality, a difference check is done on the current zones the player is inside, and the previous ones and two tables are created with this difference check: enteredZones and leftZones. All three tables (insideZones, enteredZones, and leftZones) will be nil if no zones fit them.