Add PolyZone.ensureMetatable() to fix export issue
Exports don't retain metatables, so if you return a zone from an export, the data will be returned, but none of the functions for the zone will work. PolyZone.ensureMetatable() enables you to pass in a zone and have its proper metatable set. You must have the necessary PolyZone imports for that zone type in order for this to work (i.e. if you want to call ensureMetatable on a CircleZone you must include client.lua and CircleZone.lua in your fxmanifest).