ComboZone: Add printInfo function and pzcomboinfo command
ComboZones now have the function printInfo() which will print how many zones that ComboZone has, and how many of each zone type (BoxZone, CircleZone, etc.). In addition, the command "pzcomboinfo" was added to the creation commands that will cause this printout to occur for every active ComboZone at the time of invocation.