Creation: pzcreate shortcut
When using the command "pzcreate" you can now put all the required info in the command itself to avoid having to input the info in popup text boxes. For example, with a CircleZone, instead of typing "/pzcreate circle" then filling in the popup text box for the name and radius, you can just type "/pzcreate circle circle_name 3.0" and that will start the creation of a CircleZone with the name "circle_name" and a radius of 3.0.