Implement zone events
Zone events are events attached to a zone, where if you trigger the event the callback will only be triggered if the player is in the zone.

You can trigger them from anywhere, but the most useful case is calling the server-side event "PolyZone:TriggerZoneEvent" to trigger a zone event for everyone on the server.