Refactor addEvent() to forward events rather than take callbacks
Instead of addEvent creating the event on the zone and calling a callback when the event is triggered and the player is in the zone, it instead creates a middleman event that forwards the arguments from the event trigger to the real event if the player is in the zone.

So you would call addEvent("test") and that would create an internal event that would call the "test" event if the player was in the zone when the event was triggered. That way, you can just add an event handler anywhere in the code to catch that "test" event.