Add support for targeting specific zones in a ComboZone with addEvent()
addEvent() for ComboZones takes an optional 2nd parameter of the zone name you want to target. Meaning when that zone event is triggered, instead of checking if the player is in the ComboZone, it will check if the player is in a zone inside the ComboZone, with the given name.