Refactor and improve ComboZone grid cache
After some thinking, I realized we don't need to do a circle-rectangle collision test to fill the grid cache for ComboZone. We can just grab the zone's bounding coordinates (minX, minY, maxX, maxY) and from that know every grid cell the zone should be in. This eliminates most of the prior cost of non-lazily populating the grid cache, which prevented us from doing it before.

Therefore, doing this on every ComboZone:AddZone() call does not increase the startup time since we were already adding the zone to the rows cache, and this is just one more table insert for each cell the zone should be in (usually one). The row cache has also been removed now, since it's no longer needed with this change.

I was initially worried that populating the grid non-lazily would result in a memory explosion, but this didn't happen and doesn't even make sense when I thought through it, because at most that's 1 memory pointer for each zone (which is a tiny fraction of a zone's total memory cost).

This change will make some additional features possible (when they weren't before) or much easier to implement, such as ComboZone:RemoveZone(), now that we don't have to worry about nuking the grid cache when new zones are added or existing ones are removed.

As part of the performance testing surrounding this change, I have realized that 80%+ of the cost when creating zones and adding them to a ComboZone is in the zone creation. Therefore, some more investigation is needed to determine what work is being done during zone creation and if it can be improved at all.