Add ComboZone:RemoveZone()
Improvements from the grid cache refactor have made this possible to include now. It still shouldn't be used haphazardly, but in cases where it is needed, it will be efficient enough now to use.

RemoveZone() takes a string or a function. If you pass a string, it will remove the first zone whose name matches that string. If you pass a function, it will remove the first zone where the given function returns true when the zone is passed to it.

Removing a zone will remove it from the ComboZone's internal zones table and the grid cache, but will not call destroy() on it. However, RemoveZone() will return the zone it removed (if it finds it), so you can call destroy() yourself.